API_Users/requirements.txt
flask
flask-cors
mysqlclient
python-dotenv
flask-bcrypt
flask-jwt-extended